Adds: additional properties and methods to support poll tweeting
Tweet object now has a len and str methods (just for fun). Two, nonfunctional, parameters for a short tweet and a long tweet, along with a method to return whichever is under the twitter character limit (or returns None).

The tweet_poll method in PollTweet now calls the correct length tweet string from the Tweet object and posts it, or logs an error if it received None.

Untested. Very likely broken.