Adds: basic rolling average syntax in main
Adds tweet_rolling_average function in main, which right now only goes as far as returning a dataframe of rolling averages. The next step will be to deal with creating a plot from this data and then passing it to a new PollTweet object to post to twitter.

PollParse needs a thorough once-over because right now it's about as safe as a private email server.