Adds: functionality to tweet poll averages
Added MediaTweet object to polltweet and the tweet_graph method to PollTweet. These are used similar to the Tweet object and tweet_poll method.

Added get_and_tweet_graph function to main, which handles all the data collecting, parsing, etc. Not yet added to the mainloop (need to figure out how to schedule it.)

Added additional syntax to pollparse.